+++ src/ff7/field/field.cpp
@@ -173,6 +173,22 @@ namespace ff7::field
         field_event_data_array[*ff7_externals.field_player_model_id].movement_speed = original_movement_speed;
     }
 
+    int ff7_field_load_map_trigger_data()
+    {
+        // Do not override current trigger data for woa_* fields
+        if (
+            *ff7_externals.field_resuming_from_battle_CFF268 &&
+            (
+                (*common_externals.current_field_id == 709) ||
+                (*common_externals.current_field_id == 710) ||
+                (*common_externals.current_field_id == 711)
+            )
+        )
+            return 1;
+
+        return ff7_externals.field_load_map_trigger_data_sub_6211C3();
+    }
+
     void ff7_field_hook_init()
     {
         std::copy(common_externals.execute_opcode_table, &common_externals.execute_opcode_table[OPCODE_COUNT - 1], &original_opcode_table[0]);
@@ -270,5 +286,8 @@ namespace ff7::field
         // Others fps fix
         patch_code_dword((uint32_t)&common_externals.execute_opcode_table[WAIT], (DWORD)&opcode_script_WAIT);
         replace_function(ff7_externals.sub_611BAE, opcode_IFSW_compare_sub);
+
+        // Fix wind wall animation for woa_* fields
+        replace_call_function(ff7_externals.sub_62120E + 0x3AA, ff7_field_load_map_trigger_data);
     }
 }